+++ services/frontend-service/src/ui/components/ServiceLane/ServiceLane.tsx
@@ -33,7 +33,7 @@ import { AppLockSummary, TeamLockSummary } from '../chip/EnvironmentGroupChip';
 import { WarningBoxes } from './Warnings';
 import { DotsMenu, DotsMenuButton } from './DotsMenu';
 import { useCallback, useState } from 'react';
-import { EnvSelectionDialog } from '../SelectionDialog/EnvSelectionDialog';
+import { EnvSelectionDialog } from '../SelectionDialog/SelectionDialogs';
 
 // number of releases on home. based on design
 // we could update this dynamically based on viewport width
